Revised per Jamie's feedback
