+++ app/exporters/assignment_exporter.rb
@@ -1,23 +1,16 @@
 class AssignmentExporter
-  def export_grades(assignment, students, options={})
-    CSV.generate(options) do |csv|
-      csv << headers
-      students.each do |student|
-        grade = student.grade_for_assignment(assignment)
-        grade = NullGrade.new if grade.nil? || !(grade.instructor_modified? || grade.graded_or_released?)
-        submission = student.submission_for_assignment(assignment)
-        csv << [student.first_name, student.last_name,
-                student.username, grade.score || "", grade.raw_score || "",
-                submission.try(:text_comment) || "", grade.feedback || "",
-                grade.updated_at || ""]
+  def export(course)
+    CSV.generate do |csv|
+      csv << baseline_headers
+      course.assignments.each do |assignment|
+        csv << [ assignment.id, assignment.name, assignment.point_total, assignment.description, assignment.open_at, assignment.due_at, assignment.accepts_submissions_until  ]
       end
     end
   end
 
   private
 
-  def headers
-    ["First Name", "Last Name", "Uniqname", "Score", "Raw Score",
-     "Statement", "Feedback", "Last Updated"].freeze
+  def baseline_headers
+    ["Assignment ID", "Name", "Point Total", "Description", "Open At", "Due At", "Accept Until"  ]
   end
 end
