+++ app/exporters/gradebook_exporter.rb
@@ -40,7 +40,7 @@ class GradebookExporter
     course.assignments.inject(student_data) do |memo, assignment|
       grade = assignment.grade_for_student(student)
       if grade and grade.is_student_visible?
-        memo << grade.try(:raw_score)
+        memo << grade.raw_score
       else
         memo << ''
       end
